Digital_Human_API-main/requirements.txt
basicsr==1.4.2
batch_face==1.4.0
cn2an==0.5.22
dlib==19.24.2
dominate==2.9.1
easydict==1.13
einops==0.8.0
face_alignment==1.3.5
facexlib==0.3.0
faster_whisper==1.0.1
ffmpeg_python==0.2.0
Flask==3.1.0
Flask_Cors==4.0.0
funasr==1.0.0
g2p_en==2.1.0
gdown==4.7.1
gevent==24.2.1
gradio==5.12.0
gruut==2.4.0
imageio==2.19.3
insightface==0.7.3
ipython==8.12.3
jieba_fast==0.53
kornia==0.6.8
LangSegment==0.3.2
LangSegment==0.3.5
librosa==0.9.2
lws==1.2.8
matplotlib==3.7.5
menpo==0.11.1
modelscope==1.10.0
moviepy==1.0.3
mxnet==1.9.1
nltk==3.8.1
numpy==1.23.5
onnx==1.17.0
onnxruntime==1.17.1
onnxsim==0.4.36
opencv_python==4.8.1.78
opencv_python_headless==4.9.0.80
# 可选：若需要 GPU 版 AnimeGANv2 ONNX 推理，可改为 onnxruntime-gpu 对应版本
pandas==2.0.3
Pillow==11.1.0
prettytable==3.12.0
protobuf==5.29.3
psutil==5.9.0
psutil==5.9.8
pydub==0.25.1
pyopenjtalk==0.3.3
pypinyin==0.51.0
PyQt5==5.15.11
PyQt5_sip==12.13.0
pytorch3d==0.7.4
pytorch_lightning==2.2.1
PyYAML==6.0.1
PyYAML==6.0.2
realesrgan==0.3.0
regex==2023.12.25
rembg==2.0.56
Requests==2.32.3
safetensors==0.5.2
scikit_image==0.19.3
scikit_learn==1.3.2
scipy==1.15.1
soundfile==0.12.1
tensorflow==2.18.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==1.3.2
torchvision==0.16.2+cu118
tqdm==4.66.2
transformers==4.39.3
trimesh==4.5.3
TTS==0.22.0
typeguard==4.2.1
vosk==0.3.45
whisper==1.1.10
wordsegment==1.3.1
yacs==0.1.8
huggingface_hub==0.23.5
